some very little c++11 feature demo
there might come some more code, this can be seen as a playground
testing and demonstrating features of the new c++11 standard. the
build environment is prepared for the usage of gcc with the switch
--std=c++0x in order to build properly with current versions of gcc.
by now there is one feature demo commented, since on my distribution
only gcc-4.5 series is marked stable and the mentioned feature demo
requires at least gcc-4.6. enough of talk, read the README file and
enjoy!